Split object tree and inspector components
